Doc + improvements (see commit message details)
Improvements in this commit:
 * Created bug/ & issue/ directories to provide some structure. Included a
    README file in doc/ directory.
 * Object IDs now use long/Long instead of int/Int. Injected field is private.
 * Added command line argument handling to agent for controlling stdout
    verbosity and use of the agent's network server. Cmd args are:
        [--verbose | -v]    - controls stdout verbosity
        [--server | -s]     - start a network server for event streaming.
 * Abstract classes are not instrumented. (also holds for interfaces)

+++ Dapij/src/main/java/comms/proto/CreatMsg.java
@@ -43,7 +43,7 @@ public final class CreatMsg extends Message {
             dos = new ObjectOutputStream(bos);
 
             /* Create body. */
-            dos.writeInt(creatData.getObjId());
+            dos.writeLong(creatData.getObjId());
             dos.writeObject(creatData.getClazz());
             dos.writeObject(creatData.getMethod());
             dos.writeLong(creatData.getThdId());
@@ -71,7 +71,7 @@ public final class CreatMsg extends Message {
             ObjectInputStream ois = null;
             try {
                 ois = new ObjectInputStream(bis);
-                int objId = ois.readInt();
+                long objId = ois.readLong();
                 Class<?> objCls = (Class<?>) ois.readObject();
                 String method = (String) ois.readObject();
                 long thdId = ois.readLong();
